requirements.txt
dependency-injector==4.39.1
expressintegrations==1.3.7
fastapi==0.78.0
google-cloud-firestore==2.5.2
google-cloud-logging==3.1.1
google-cloud-secret-manager==2.11.0
google-cloud-storage==2.3.0
google-cloud-tasks==2.9.0
googleapis-common-protos==1.56.2
gunicorn==20.1.0
pydantic==1.9.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
urllib3==1.26.9
uvicorn==0.17.6